requirements.txt
uv
streamlit==1.38.0
python-dotenv==1.0.1
pydantic==2.8.2
langchain==0.2.14
langchain-core==0.2.38
langchain-community==0.2.12
openai==1.43.0
google-generativeai==0.7.2